chore: reorder helpwave icons
