+++ landing-page/components/Navigation.tsx
@@ -1,4 +1,5 @@
 import { Span } from '@helpwave/common/components/Span'
+import { Helpwave } from '@helpwave/common/icons/Helpwave'
 import { tw } from '@helpwave/common/twind'
 import Link from 'next/link'
 
@@ -14,10 +15,11 @@ const Navigation = () => {
   return (
     <div className={tw('max-w-screen fixed left-1/2 top-[40px] -translate-x-1/2')}>
       <div className={tw('flex flex-wrap gap-8 items-center justify-between mx-auto p-4')}>
+        <Helpwave />
         {items.map(({ name, url }) => (
           <div key={url} className={tw('text-xl group')}>
             <Link href={url}>
-              <Span type="subsectionTitle">
+              <Span type="navigationItem">
                 {name}
               </Span>
             </Link>
